+++ lib/vwo.rb
@@ -1,4 +1,4 @@
-# Copyright 2025 Wingify Software Pvt. Ltd.
+# Copyright 2024-2025 Wingify Software Pvt. Ltd.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -36,6 +36,7 @@ class VWO
                    .set_network_manager
                    .set_segmentation
                    .init_polling
+                   .init_batch
 
     if options[:settings]
       return @@vwo_builder.build(options[:settings])