+++ lib/vwo/packages/storage/storage.rb
@@ -1,4 +1,4 @@
-# Copyright 2025 Wingify Software Pvt. Ltd.
+# Copyright 2024-2025 Wingify Software Pvt. Ltd.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,9 +18,11 @@ class Storage
   @instance = nil
 
   attr_reader :connector
+  attr_accessor :is_storage_enabled
 
   def initialize
     @connector = nil
+    @is_storage_enabled = false
   end
 
   # Attach a connector (can be an instance or a class)