+++ lib/vwo/vwo_builder.rb
@@ -1,4 +1,4 @@
-# Copyright 2025 Wingify Software Pvt. Ltd.
+# Copyright 2024-2025 Wingify Software Pvt. Ltd.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -21,9 +21,13 @@ require_relative './packages/storage/storage'
 require_relative './packages/network_layer/manager/network_manager'
 require_relative './packages/segmentation_evaluator/core/segmentation_manager'
 require_relative './services/logger_service'
+require_relative './services/batch_event_queue'
+require_relative './utils/function_util'
+require_relative './utils/batch_event_dispatcher'
+require_relative './constants/constants'
 
 class VWOBuilder
-  attr_reader :settings, :storage, :log_manager, :is_settings_fetch_in_progress, :vwo_instance
+  attr_reader :settings, :storage, :log_manager, :is_settings_fetch_in_progress, :vwo_instance, :is_valid_poll_interval_passed_from_init
 
   # Initialize the VWOBuilder with the given options
   # @param options [Hash] The options for the VWOBuilder
@@ -33,7 +37,48 @@ class VWOBuilder
     @storage = nil
     @log_manager = nil
     @is_settings_fetch_in_progress = false
+    @is_valid_poll_interval_passed_from_init = false
     @vwo_instance = nil
+  end
+
+  # Initializes the batch event processing system
+  # Validates batch event settings and configures the BatchEventsQueue
+  # Sets up event dispatcher and flushes any existing events
+  # @raise [StandardError] If batch event configuration is invalid
+  def init_batch
+    # if gateway service is configured, then do not initialize batch event queue
+    if SettingsService.instance.is_gateway_service_provided
+      LoggerService.log(LogLevelEnum::INFO, "GATEWAY_AND_BATCH_EVENTS_CONFIG_MISMATCH")
+      return self
+    end
+    begin
+      if @options.key?(:batch_event_data)
+        if @options[:batch_event_data].is_a?(Hash)
+          # Validate batch event parameters
+          events_per_request = @options[:batch_event_data][:events_per_request]
+          request_time_interval = @options[:batch_event_data][:request_time_interval]
+          
+          if (!events_per_request.is_a?(Numeric) || events_per_request <= 0) &&
+             (!request_time_interval.is_a?(Numeric) || request_time_interval <= 0)
+            LoggerService.log(LogLevelEnum::ERROR, "INVALID_BATCH_EVENTS_CONFIG")
+          end
+          
+          BatchEventsQueue.configure(
+            @options[:batch_event_data].merge(
+              {
+                account_id: @options[:account_id],
+                dispatcher: method(:dispatcher)
+              }
+            )
+          )
+          @batch_event_data = @options[:batch_event_data]
+        else
+          LoggerService.log(LogLevelEnum::ERROR, "Invalid batch events config, should be a hash.", nil)
+        end
+      end
+    rescue StandardError => e
+      LoggerService.log(LogLevelEnum::ERROR, "Failed to initialize batch event queue: #{e.message}", nil)
+    end
   end
 
   # Set the network manager
@@ -89,7 +134,12 @@ class VWOBuilder
   # Set the storage
   # @return [VWOBuilder] The VWOBuilder instance
   def set_storage
-    @storage = @options[:storage] ? Storage.instance.attach_connector(@options[:storage]) : nil
+    if @options[:storage]
+      @storage = Storage.instance.attach_connector(@options[:storage])
+      Storage.instance.is_storage_enabled = true
+    else
+      @storage = nil
+    end
     LoggerService.log(LogLevelEnum::DEBUG, "SERVICE_INITIALIZED", {service: "Storage"})
     self
   end
@@ -116,26 +166,19 @@ class VWOBuilder
   # Initialize the polling
   # @return [VWOBuilder] The VWOBuilder instance
   def init_polling
-    return self unless @options[:poll_interval]
-
-    unless @options[:poll_interval].is_a?(Numeric)
+    poll_interval = @options[:poll_interval]
+    
+    if poll_interval && poll_interval.is_a?(Numeric) && poll_interval >= 1000
+      # this is to check if the poll_interval passed in options is valid
+      @is_valid_poll_interval_passed_from_init = true
+      check_and_poll
+    elsif poll_interval
+      # only log error if poll_interval is present in options
       LoggerService.log(LogLevelEnum::ERROR, "INIT_OPTIONS_INVALID", {
         key: 'poll_interval',
-        correctType: 'number'
+        correctType: 'number >= 1000'
       })
-      return self
     end
-
-    # Check if the polling interval is greater than or equal to 1000
-    unless @options[:poll_interval] >= 1000
-      LoggerService.log(LogLevelEnum::ERROR, "INIT_OPTIONS_INVALID", {
-        key: 'poll_interval',
-        correctType: 'number'
-      })
-      return self
-    end
-
-    check_and_poll
     self
   end
 
@@ -144,25 +187,42 @@ class VWOBuilder
   # @return [VWOClient] The VWO instance
   def build(settings)
     @vwo_instance = VWOClient.new(settings, @options)
+    # if poll_interval is not present in options, set it to the pollInterval from settings
+    update_poll_interval_and_check_and_poll(settings)
     @vwo_instance
+  end
+
+  def update_poll_interval_and_check_and_poll(settings, should_check_and_poll = true)
+    # only update the poll_interval if it poll_interval is not valid or not present in options
+    if !@is_valid_poll_interval_passed_from_init
+      @options[:poll_interval] = settings["pollInterval"] || Constants::POLLING_INTERVAL
+      LoggerService.log(LogLevelEnum::DEBUG, "USING_POLL_INTERVAL_FROM_SETTINGS", {
+        source: settings["pollInterval"] ? 'settings' : 'default',
+        pollInterval: @options[:poll_interval]
+      })
+    end
+    # should_check_and_poll will be true only when we are updating the poll_interval first time from self.build method
+    # if we are updating the poll_interval already running polling, we don't need to check and poll again
+    if should_check_and_poll && !@is_valid_poll_interval_passed_from_init
+      check_and_poll
+    end
   end
 
   # This method is used to check and poll the settings from the server
   # @return [VWOBuilder] The VWOBuilder instance
   def check_and_poll
-    polling_interval = @options[:poll_interval]
-
     @thread_pool = NetworkManager.instance.get_client.get_thread_pool
     @thread_pool.post do
       loop do
-        sleep(polling_interval / 1000.0)
+        sleep(@options[:poll_interval]/ 1000.0)
         begin
           latest_settings = fetch_settings(true)
-          if latest_settings.to_json != @settings.to_json
+          if latest_settings && latest_settings.to_json != @settings.to_json
             @settings = latest_settings
             LoggerService.log(LogLevelEnum::INFO, "POLLING_SET_SETTINGS")
             @vwo_instance.update_settings(latest_settings.clone, false) if @vwo_instance
-          else
+            update_poll_interval_and_check_and_poll(latest_settings, false)
+          elsif latest_settings
             LoggerService.log(LogLevelEnum::INFO, "POLLING_NO_CHANGE_IN_SETTINGS")
           end
         rescue StandardError => e
@@ -170,5 +230,18 @@ class VWOBuilder
         end
       end
     end
+  end
+
+  def dispatcher(events, callback)
+    BatchEventDispatcher.dispatch(
+      {
+        ev: events
+      },
+      callback,
+      {
+        a: @options[:account_id],
+        env: @options[:sdk_key]
+      }
+    )
   end
 end